* V5-7-patches:
  DISMAN EXPRESSION MIB: Add a regression test
  new DISMAN EXPRESSION MIB: Report success when destroying a non-existing row
  new DISMAN EXPRESSION MIB: Fix two compiler warnings
  old DISMAN EXPRESSION MIB: Disable recursion
  old DISMAN EXPRESSION MIB: Use callback transport instead of UDP over localhost
  old DISMAN EXPRESSION MIB: Move common code into a new function
  old DISMAN EXPRESSION MIB: Simplify Evaluate_Expression()
  old DISMAN EXPRESSION MIB: Add more logging
  old DISMAN EXPRESSION MIB: Fix row destruction code
  old DISMAN EXPRESSION MIB: Fix another use-after-free
  old DISMAN EXPRESSION MIB: Convert exit() calls into return statements
  old DISMAN EXPRESSION MIB: Fix expression evaluator
  old DISMAN EXPRESSION MIB: Declare several variables as const
  old DISMAN EXPRESSION MIB: Improve source code readability
  old DISMAN EXPRESSION MIB: Change two global variables into local variables
  old DISMAN EXPRESSION MIB: Declare local functions and variables static
  old DISMAN EXPRESSION MIB: Fix a use-after-free
  old DISMAN EXPRESSION MIB: Fix two memory leaks in expValueTable_add()
  old DISMAN EXPRESSION MIB: Fix a grammatical error in a variable name
  old DISMAN EXPRESSION MIB: Fix a spelling error in a variable name

+++ agent/mibgroup/disman/expr/expValue.c
@@ -376,7 +376,7 @@ _expValue_evalExpr(  netsnmp_variable_list *expIdx,
                                       suffix, suffix_len );
             if ( var && var->type == ASN_NULL ) {
                 /* Adjust position of failure */
-                var->data = (void *)(cp1 - exprRaw + (int)var->data);
+                var->data = (void *)(cp1 - exprRaw + (uintptr_t)var->data);
                 return var;
             } else if (*cp2 != ')') {
                 snmp_free_var(exprAlDente);
@@ -814,7 +814,7 @@ _expValue_setError( struct expExpression *exp, int reason,
         return;
     exp->expErrorCount++;
  /* exp->expErrorTime  = NOW; */
-    exp->expErrorIndex = ( var && var->data ? (int)var->data : 0 );
+    exp->expErrorIndex = ( var && var->data ? (uintptr_t)var->data : 0 );
     exp->expErrorCode  = reason;
     memset( exp->expErrorInstance, 0, sizeof(exp->expErrorInstance));
     memcpy( exp->expErrorInstance, suffix, suffix_len * sizeof(oid));
